+++ src/gui/update.rs
@@ -22,6 +22,7 @@ use {
         repository::{RepositoryKind, RepositoryPackage},
         utility::{download_update_to_temp_file, wow_path_resolution},
     },
+    ajour_widgets::header::ResizeEvent,
     anyhow::Context,
     async_std::sync::{Arc, Mutex},
     chrono::{NaiveTime, Utc},
@@ -32,7 +33,6 @@ use {
     std::convert::TryFrom,
     std::hash::Hasher,
     std::path::{Path, PathBuf},
-    widgets::header::ResizeEvent,
 };
 
 pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Message>> {
@@ -168,7 +168,7 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                     if addon.is_updatable(&package) {
                         addon.state = AddonState::Updatable;
                     } else {
-                        addon.state = AddonState::Ajour(None);
+                        addon.state = AddonState::Idle;
                     }
                 }
             };
@@ -457,7 +457,7 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
 
                             // Check if addon is updatable based on release channel.
                             if let Some(package) = a.relevant_release_package() {
-                                if a.is_updatable(&package) && a.state != AddonState::Corrupted {
+                                if a.is_updatable(&package) {
                                     a.state = AddonState::Updatable;
                                 }
                             }
@@ -524,10 +524,20 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                     log_error(&error);
                     ajour.error = Some(error);
 
-                    if reason == DownloadReason::Install {
-                        if let Some(install_addon) = install_addons.iter_mut().find(|a| a.id == id)
-                        {
-                            install_addon.status = InstallStatus::Retry;
+                    match reason {
+                        DownloadReason::Update => {
+                            if let Some(_addon) =
+                                addons.iter_mut().find(|a| a.primary_folder_id == id)
+                            {
+                                _addon.state = AddonState::Retry;
+                            }
+                        }
+                        DownloadReason::Install => {
+                            if let Some(install_addon) =
+                                install_addons.iter_mut().find(|a| a.id == id)
+                            {
+                                install_addon.status = InstallStatus::Retry;
+                            }
                         }
                     }
                 }
@@ -608,10 +618,20 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                     log_error(&error);
                     ajour.error = Some(error);
 
-                    if reason == DownloadReason::Install {
-                        if let Some(install_addon) = install_addons.iter_mut().find(|a| a.id == id)
-                        {
-                            install_addon.status = InstallStatus::Retry;
+                    match reason {
+                        DownloadReason::Update => {
+                            if let Some(_addon) =
+                                addons.iter_mut().find(|a| a.primary_folder_id == id)
+                            {
+                                _addon.state = AddonState::Retry;
+                            }
+                        }
+                        DownloadReason::Install => {
+                            if let Some(install_addon) =
+                                install_addons.iter_mut().find(|a| a.id == id)
+                            {
+                                install_addon.status = InstallStatus::Retry;
+                            }
                         }
                     }
                 }
@@ -693,9 +713,9 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
             let addons = ajour.addons.entry(flavor).or_default();
             if let Some(addon) = addons.iter_mut().find(|a| a.primary_folder_id == id) {
                 if result.is_ok() {
-                    addon.state = AddonState::Ajour(Some("Completed".to_owned()));
+                    addon.state = AddonState::Completed;
                 } else {
-                    addon.state = AddonState::Ajour(Some("Error".to_owned()));
+                    addon.state = AddonState::Error("Error".to_owned());
                 }
             }
         }
@@ -802,7 +822,7 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                         if addon.is_updatable(&package) {
                             addon.state = AddonState::Updatable;
                         } else {
-                            addon.state = AddonState::Ajour(None);
+                            addon.state = AddonState::Idle;
                         }
                     }
 
@@ -1465,7 +1485,7 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
             if let Some(last_updated) = &ajour.catalog_last_updated {
                 let now = Utc::now();
                 let now_time = now.time();
-                let refresh_time = NaiveTime::from_hms(0, 5, 0);
+                let refresh_time = NaiveTime::from_hms(0, 40, 0);
 
                 if last_updated.date() < now.date() && now_time > refresh_time {
                     log::debug!("Message::RefreshCatalog: catalog needs to be refreshed");